requirements.txt
tensorflow==2.18.0
numpy==2.0.2
pillow==11.3.0
streamlit==1.47.0



